fix(.editorconfig):  make C files save w/o BOM
Since my development workstation is on a Windows platform,
when I make modifications/fixes to LVGL library code, I often
like to do it in this project first because it offers a great place
to test the results.  I found that when I did that and then copied
the changes to the LVGL project fork I have, they were coming
with BOMs in the files, and the LVGL repository maintainers are
trying to keep the BOMs out of the source files because they cause
problems in some places.  I offer this as help, since most of the
C source files under these projects is from the LVGL repository.