fix(filters):  fix clean up of filter files incomplete
"Filter records" are Visual Studio's name for "folders", and this
project cleverly creates these folders to match the file LVGL pjt
directory structure.

The problem:

LvglProjectFileUpdater project has 3 places where .vcxproj.filters files
are loaded, records removed before a new set of filter records are added.
However, for all 3 projects that it maintains, it is not removing all the
filter records first as intended. The result of this is that the records
that are not removed remain in the file and accumulate when the new
filter records are added.

Resolves #91